Move Logo component into subfolder
